Update publiccode.yml
update publiccode for
Message
publiccode.yml:5	publiccodeYmlVersion: v0.2 is not the latest version, use '0.4.0'. Parsing this file as v0.4.0.
publiccode.yml:26	description.it.genericName: This key is DEPRECATED and will be removed in the future
publiccode.yml:206	description.it.documentation: 'https://www.latraccia.it/prodotti/sic/%20%20%20' not reachable: HTTP GET failed for https://www.latraccia.it/prodotti/sic/%20%20%20: forbidden resource